+++ app/components/chat/ChatRoomList.css
@@ -15,5 +15,4 @@
   border-radius: 50%;
   font-size: 12px;
   color: #666;
-
 }
