+++ app/layout.tsx
@@ -3,6 +3,8 @@ import { ThemeModeScript } from "flowbite-react";
 import "./globals.css";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
+import Script from "next/script";
+import KakaoChat from "./components/chat/KakaoChat";
 
 export const metadata: Metadata = {
   title: "paranmanzang",
@@ -26,7 +28,16 @@ export default function RootLayout({
           {children}
         </main>
         <Footer />
+        <KakaoChat />
       </body>
+       {/* Kakao SDK 스크립트 로드 */}
+       {/* <Script
+            src="https://t1.kakaocdn.net/kakao_js_sdk/2.7.2/kakao.min.js"
+            integrity="sha384-TiCUE00h649CAMonG018J2ujOgDKW/kVWlChEuu4jK2vxfAAD0eZxzCKakxg55G4"
+            crossOrigin="anonymous"
+            strategy="afterInteractive" // 클라이언트 사이드에서만 로드되도록 설정
+          />
+          */}
     </html>
   );
 }
