--- app/model/test.ts
@@ -1,1 +1,0 @@
-// 여기에 데이터에서 받아오는 정제 값들 이나 interface 활용할 것 빼두기
